+++ webengineviewer/src/checkphishingurl/checkphishingurlutil.h
@@ -31,7 +31,7 @@ QString versionApps();
 QString databaseFileName();
 WEBENGINEVIEWER_EXPORT quint16 minorVersion();
 WEBENGINEVIEWER_EXPORT quint16 majorVersion();
-WEBENGINEVIEWER_EXPORT int convertToSecond(QString str);
+WEBENGINEVIEWER_EXPORT double convertToSecond(QString str);
 }
 }
 #endif // CHECKPHISHINGURLUTIL_H
